OpenAPI spec updated and used for encrypt status.
[ upstream commit 7cdadbcea9c9f48e717a670ab287b1933703705a ]

Global variable `countErrors` converted to the function local.

+++ cilium-dbg/cmd/encrypt_status_test.go
@@ -16,7 +16,7 @@ const procTestFixtures = "fixtures/proc"
 func (s *EncryptStatusSuite) TestGetXfrmStats(c *C) {
 	errCount, m, err := getXfrmStats(procTestFixtures)
 	c.Assert(err, Equals, nil)
-	currentCount := 0
+	currentCount := int64(0)
 	testCases := []struct {
 		name string
 		want int64
